requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cpip==0.9.7
cryptography==36.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.1
distro==1.6.0
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.46.0
django-appconf==1.0.5
django-avatar==5.0.0
django-comments-xtd==2.9.4
django-contrib-comments==2.1.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-markdown2==0.3.1
django-payments==0.15.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
docutils==0.17.1
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
longerusername==0.4
markdown2==2.4.2
myapp==0.1.dev0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
netaddr==0.8.0
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.2
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
python-gettext==4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
South==1.0.2
sqlparse==0.4.1
urllib3==1.26.5
var-dump==1.2
virtualenv==20.4.3
vk-api==11.9.4
whitenoise==5.3.0
wrapt==1.13.3
yookassa==2.2.3

{"type":"notification","event":"payment.succeeded","object":{"id":"2203aa1d-000f-5000-8000-17102541fd31",
"status":"succeeded","paid":true,"amount":{"value":"1.00","currency":"RUB"},"captured_at":"2018-01-31T10:12:06.249Z",
"created_at":"2018-01-31T10:11:41.499Z","description":"Оплата тестового заказа №100500 для магазина Вереница","metadata":{"zakaz":"100500","param3":"value3","param2":"value2"},
"payment_method":{"type":"bank_card","id":"2203aa1d-000f-5000-8000-17102541fd31","saved":false,
"card":{"last4":"1026","expiry_month":"12","expiry_year":"2025","card_type":"Unknown"},"title":"Bank card *1026"},
"recipient":{"account_id":"500105","gateway_id":"1500105"},"refunded_amount":{"value":"0.00","currency":"RUB"},"test":true}}